Export FConfigurationDictionary
