+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@freemework/common",
-	"version": "0.8.4",
+	"version": "0.8.5",
 	"description": "Common library of the Freemework Project.",
 	"author": {
 		"name": "Freemework Project",
