+++ src/configuration/FConfiguration.ts
@@ -1,6 +1,8 @@
 import { FExceptionArgument } from "../exception";
-import { FConfigurationException } from "./FConfigurationException";
+import { FUnreadonly } from "../util";
+
 import { FConfigurationValue } from "./FConfigurationValue";
+
 
 /**
  * The `FConfiguration` provides contract to access key-value configuration sources
@@ -82,7 +84,7 @@ export abstract class FConfiguration {
 		const jsonDict: { [key: string]: boolean | number | string; } = {};
 		expandJson(jsonObject, jsonDict);
 
-		const dict: Unreadonly<FConfiguration.Dictionary> = {};
+		const dict: FUnreadonly<FConfigurationDictionary.Data> = {};
 		for (const [name, value] of Object.entries(jsonDict)) {
 			if (typeof value === "string") {
 				dict[name] = value === "" ? null : value;
@@ -96,7 +98,7 @@ export abstract class FConfiguration {
 		const encodedJson: string = encodeURIComponent(JSON.stringify(jsonObject));
 		const sourceURI: string = `configuration:json?data=${encodedJson}`;
 
-		return new _FConfiguration(
+		return new FConfigurationDictionary(
 			new URL(sourceURI),
 			Object.freeze(dict)
 		);
@@ -104,6 +106,7 @@ export abstract class FConfiguration {
 
 
 	protected static readonly DEFAULT_INDEXES_KEY = "indexes";
+
 
 	/**
 	 * Gets configuration source URI.
@@ -288,138 +291,6 @@ export abstract class FConfiguration {
 	public abstract has(key: string): boolean;
 }
 
-export namespace FConfiguration {
-	export type Dictionary = { readonly [key: string]: string | null };
-}
+// Import here due to cyclic dependencies
 
-
-class _FConfiguration extends FConfiguration {
-
-	private readonly _dict: FConfiguration.Dictionary;
-	private readonly _sourceURI: URL;
-	private readonly _configurationNamespace: string | null;
-	private _keys: ReadonlyArray<string> | null;
-
-	public constructor(sourceURI: URL, dict: FConfiguration.Dictionary, configurationNamespace?: string) {
-		super();
-		this._dict = Object.freeze({ ...dict });
-		this._sourceURI = sourceURI;
-		this._configurationNamespace = configurationNamespace !== undefined ? configurationNamespace : null;
-		this._keys = null;
-	}
-
-	public get configurationNamespace(): string | null {
-		return this._configurationNamespace;
-	}
-
-	public get keys(): ReadonlyArray<string> {
-		return this._keys !== null ? this._keys : (this._keys = Object.freeze(Object.keys(this._dict)));
-	}
-
-	public get sourceURI(): URL {
-		return this._sourceURI;
-	}
-
-	public findNamespace(configurationNamespace: string): FConfiguration | null {
-		throw new Error("Method not implemented.");
-	}
-
-	public find(key: string): FConfigurationValue | null {
-		if (key in this._dict) {
-			const valueData = this._dict[key];
-			const value: FConfigurationValue = FConfigurationValue.factory(
-				key,
-				valueData,
-				this.sourceURI,
-				null
-			);
-			return value;
-		} else {
-			return null;
-		}
-	}
-
-	public getNamespace(configurationNamespace: string): FConfiguration {
-		const innerDict: Unreadonly<FConfiguration.Dictionary> = {};
-		const criteria = configurationNamespace + ".";
-		const criteriaLen = criteria.length;
-		Object.keys(this._dict).forEach((key) => {
-			if (key.length > criteriaLen && key.startsWith(criteria)) {
-				const value = this._dict[key];
-				innerDict[key.substring(criteriaLen)] = value;
-			}
-		});
-
-		const innerConfigurationNamespace = this._configurationNamespace !== null ?
-			`${this._configurationNamespace}.${configurationNamespace}` : configurationNamespace;
-
-		if (Object.keys(innerDict).length === 0) {
-			throw new FConfigurationException(
-				`Namespace '${innerConfigurationNamespace}' was not found in the configuration.`,
-				innerConfigurationNamespace
-			);
-		}
-		return new _FConfiguration(this.sourceURI, innerDict, innerConfigurationNamespace);
-	}
-
-	public get(key: string, defaultData?: string | null): FConfigurationValue {
-		if (key in this._dict) {
-			let valueData = this._dict[key];
-
-			if (valueData === null && defaultData !== undefined) {
-				valueData = defaultData;
-			}
-
-			const value: FConfigurationValue = FConfigurationValue.factory(
-				key, valueData, this.sourceURI, null
-			);
-			return value;
-		} else if (defaultData !== undefined) {
-			const value: FConfigurationValue = FConfigurationValue.factory(
-				key, defaultData, this.sourceURI, null
-			);
-			return value;
-		} else {
-			const value: FConfigurationValue = FConfigurationValue.factoryUndefined(
-				key
-			);
-			return value;
-		}
-	}
-
-	public getArray(key: string, indexesName: string = FConfiguration.DEFAULT_INDEXES_KEY): Array<FConfiguration> {
-		const arrayIndexesKey = `${key}.${indexesName}`;
-		const arrayIndexes: Array<string> = this.get(arrayIndexesKey).asString
-			.split(" ")
-			.filter(s => s !== "");
-
-		const arrayNamespaces: Array<FConfiguration> = arrayIndexes.map(s => {
-			const arrayItemNamespaceKey = `${key}.${s}`;
-			return this.getNamespace(arrayItemNamespaceKey);
-		});
-
-		return arrayNamespaces;
-	}
-
-	public hasNamespace(configurationNamespace: string): boolean {
-		const criteria = configurationNamespace + ".";
-		const criteriaLen = criteria.length;
-		for (const key of Object.keys(this._dict)) {
-			if (key.length > criteriaLen && key.startsWith(criteria)) {
-				return true;
-			}
-		}
-		return false;
-	}
-
-	public has(key: string): boolean {
-		return key in this._dict;
-	}
-}
-
-/**
- * Remove readonly modifier for all properties in T. Opposite util for `Readonly<T>`.
- */
-type Unreadonly<T> = {
-	-readonly [P in keyof T]: T[P];
-};
+import { FConfigurationDictionary } from "./FConfigurationDictionary";
