+++ src/configuration/index.ts
@@ -1,4 +1,5 @@
 export { FConfiguration } from "./FConfiguration";
 export { FConfigurationChain } from "./FConfigurationChain";
+export { FConfigurationDictionary } from "./FConfigurationDictionary";
 export { FConfigurationException } from "./FConfigurationException";
 export { FConfigurationValue } from "./FConfigurationValue";
