+++ src/util/FUnreadonly.ts
@@ -1,0 +1,6 @@
+/**
+ * Remove readonly modifier for all properties in T. Opposite util for `Readonly<T>`.
+ */
+export type FUnreadonly<T> = {
+	-readonly [P in keyof T]: T[P];
+};
